(BSR)[API] refactor: move create_products_thumb to utils file
